Extract tooling functions to their own class
